# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Application for planning your journey",
        "title": "Journer Planner",
        "contact": {}
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/groups/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create new group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "AddGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of group",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "by adding logins you will automatically invite this users",
                        "name": "invites",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/groups/declineinvite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Decline invite",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "Decline invite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of invite",
                        "name": "invite_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/groups/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete group by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "DeleteGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of group",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/groups/getgroupinfo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get full info about group you are a member of",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "GetGroupInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of group",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/groups/getlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all the groups you are a member of",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "GetGroups",
                "responses": {}
            }
        },
        "/groups/givelead": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Give another member of group leader role",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "GiveLeaderRole",
                "parameters": [
                    {
                        "type": "string",
                        "description": "member login",
                        "name": "user_login",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of group",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/groups/invite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Invite user to group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "Invite user to group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of group",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "invited user",
                        "name": "user_login",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/groups/invitelist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get your list of invites",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "invites"
                ],
                "summary": "Get invite list",
                "responses": {}
            }
        },
        "/groups/leaveGroup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Leave from group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "LeaveFromGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of group",
                        "name": "group_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}